Bugfix/typeFix : 오브젝트 좌표 값 데이터 타입 변경

+++ src/main/java/com/mtvs/arzip/domain/dto/ar_space_data/ARObjectPlacementDataResponse.java
@@ -11,10 +11,10 @@ import lombok.*;
 public class ARObjectPlacementDataResponse {
 
     private Long no;
-    private Integer x;
-    private Integer y;
-    private Integer z;
-    private Integer rotation;
+    private double x;
+    private double y;
+    private double z;
+    private double rotation;
     private Long arSpaceDataNo;
     private Long objectInfoNo;
