+++ src/main/java/com/mtvs/arzip/domain/entity/ARObjectPlacementData.java
@@ -16,10 +16,10 @@ public class ARObjectPlacementData {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long no;
 
-    private Integer x;  // x 좌표 값
-    private Integer y;  // y 좌표 값
-    private Integer z;  // z 좌표 값
-    private Integer rotation; // 오브젝트 회전 값
+    private double x;  // x 좌표 값
+    private double y;  // y 좌표 값
+    private double z;  // z 좌표 값
+    private double rotation; // 오브젝트 회전 값
 
     @JoinColumn(name = "arSpaceData_no")
     @ManyToOne(fetch = FetchType.EAGER)
